requirements.txt
transformers
datasets
peft
trl
bitsandbytes
wandb
python-dotenv
